requirements.txt
Django==5.1.7
html5lib==1.1
pillow==11.1.0
reportlab==4.3.1
xhtml2pdf==0.2.17
django-widget-tweaks==1.5.0
cssselect2==0.8.0
whitenoise==6.9.0
Brotli==1.1.0
asgiref==3.8.1
tzdata==2025.2
sqlparse==0.5.3
psycopg2-binary==2.9.10
dj-database-url==3.0.0
gunicorn==23.0.0
uvicorn==0.34.3
django-import-export==4.3.8
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
python-dotenv==1.1.1